WIP missings implementation InitializeDatabase unit test

+++ Infrastructure/Interfaces/IExtensionDatabaseMetadata.cs
@@ -53,8 +53,9 @@ namespace Infrastructure.Interfaces
 
         /// <summary>
         /// The permission levels that will be created in application (only by Security extension).
+        /// Id is forced by design
         /// </summary>
-        IEnumerable<Tuple<int, int, string, string>> PermissionLevelIdValueLabelAndTips { get; }
+        IEnumerable<Tuple<int, byte, string, string>> PermissionLevelIdValueLabelAndTips { get; }
 
         /// <summary>
         /// The users you wish to add to the application.
